Create Web_Scrapper.py
This code specifies two functions: 'scrape_verge' and 'save_csv', which scrape the website and save the data to a CSV file, and 'save_db', which scan the website and save the data to a SQLite database. The code uses these methods in the main block to scrape the website, save the data to a CSV file and a SQLite database, and deduplicate the data in the database to prevent repetitions.

Note: You will need to execute this code on a cloud service such as AWS on a daily basis in order to save the articles in the database.